refactor(app): decompose SettingsScreen.tsx into settings/ section components
Extract the inlined notification-prefs, voice/input, and biometric-lock
sections out of the ~840-LOC SettingsScreen body (#5655). Behaviour-
preserving, structure-only: same toggles, handlers, conditional rendering,
testIDs, accessibility props, and the notification-pref Switch WS-closed
accessibilityRole="alert" failure path.

New components under packages/app/src/components/settings/:
- NotificationPrefsSection — categories + per-device mutes + quiet-hours +
  known-devices wiring (#4542/#4543/#4544/#4564)
- VoiceInputSection — chat/terminal enter-to-send, speech language + voice
  input mode rows and picker sheets; owns its local picker/speech-lang state
- SecuritySection — biometric-lock toggle
- QuietHoursEditor / KnownDevicesList — relocated from the screen body so the
  new section can compose them
- styles.ts / constants.ts — shared StyleSheet + label/order/message/option
  constants consumed by the screen and the sections

State stays in SettingsScreen and is passed down as props (mirroring the
existing QuietHoursEditor/KnownDevicesList prop pattern); section-local UI
state (picker visibility, quiet-hours draft) stays local.

The settings static-source tests now read SettingsScreen.tsx plus every
extracted settings component and search the combined source, so all existing
assertions hold without weakening them.

SettingsScreen.tsx: 1683 -> 601 LOC. Full app suite (1865 tests) + tsc green.
